requirements/base.txt
Django>=1.11,<1.12
gunicorn
django-dotenv
dj-database-url
psycopg2
raven
whitenoise
ipython
markdown
django-compressor
libsass
lxml
pypandoc
PyYAML
python-dateutil
